Missed moving ValidCompactorTxnList to storage-api.  Fixing it.
Small formatting changes.
