+++ storage-api/src/java/org/apache/hadoop/hive/common/ValidCompactorTxnList.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -64,7 +64,7 @@ public class ValidCompactorTxnList extends ValidReadTxnList {
     else {
       lastElementPos = idx;
     }
-    /**
+    /*
      * ensure that we throw out any exceptions above highWatermark to make
      * {@link #isTxnValid(long)} faster 
      */
